Moving models to the models folder
