+++ package.json
@@ -20,6 +20,7 @@
     "@angular/platform-browser-dynamic": "^17.0.6",
     "@angular/router": "^17.0.6",
     "@angular/service-worker": "17.0.6",
+    "@ngrx/signals": "^17.2.0",
     "body-parser": "^1.20.3",
     "bootstrap": "^3.4.1",
     "codelyzer": "^6.0.2",
