+++ src/app/current-conditions/current-conditions.component.ts
@@ -2,7 +2,7 @@ import {Component, inject, Signal} from '@angular/core';
 import {WeatherService} from '../weather.service';
 import {LocationService} from '../location.service';
 import {Router} from '@angular/router';
-import {ConditionsAndZip} from '../conditions-and-zip.type';
+import {ConditionsAndZip} from '../model/conditions-and-zip.type';
 
 @Component({
   selector: 'app-current-conditions',
