+++ src/app/model/current-conditions.type.ts
