+++ src/app/weather.service.ts
@@ -2,8 +2,8 @@ import {Injectable, Signal, signal} from '@angular/core';
 import {Observable} from 'rxjs';
 
 import {HttpClient} from '@angular/common/http';
-import {CurrentConditions} from './current-conditions/current-conditions.type';
-import {ConditionsAndZip} from './conditions-and-zip.type';
+import {CurrentConditions} from './model/current-conditions.type';
+import {ConditionsAndZip} from './model/conditions-and-zip.type';
 import {Forecast} from './forecasts-list/forecast.type';
 
 @Injectable()
